fix(tests): move AIBuilderPage to components dir (spans multiple UI areas)
AIBuilderPage interacts with both canvas-choice-prompt and
workflow-suggestions which are separate, unrelated DOM trees.
Scoping to a single container broke the workflow-builder E2E tests.
Moving to components/ exempts it from scope-lockdown.
